fix: create separate util folder for query function and query mutation
