+++ client/src/constants/index.jsx
@@ -1,3 +1,4 @@
 const BASE_URL = "http://localhost:8000/api/todo";
+const PRODUCTS_URL = "https://dummyjson.com/products";
 
-export { BASE_URL };
+export { BASE_URL, PRODUCTS_URL };
